refactor: update IPC handling and remove unused serde utility
